requirements.txt
tweepy==3.9.0
pyyaml==5.3.1
pyshorteners==1.0.1